Introducing a shared error maper
